# New gofast protobuf code