--- a/generate.go
+++ b/generate.go
@@ -1,6 +1,6 @@
 package veneur
 
-//go:generate protoc --go_out=. ssf/sample.proto
+//go:generate protoc --gofast_out=. ssf/sample.proto
 //go:generate gojson -input config_spec.yaml -o config.go -fmt yaml -pkg veneur -name Config
 //go:generate gojson -input example_proxy.yaml -o config_proxy.go -fmt yaml -pkg veneur -name ProxyConfig
 //TODO(aditya) reenable go:generate gojson -input fixtures/datadog_trace.json -o datadog_trace_span.go -fmt json -pkg veneur -name DatadogTraceSpan