… ssf/sample.pb.go