#### Summary
Uses [gogo-protobuf](https://github.com/gogo/protobuf)'s `gofast` code generator for faster marshal and unmarshaling.

#### Motivation
I stumbled into a comment somewhere that gogo-protobuf was way faster than stock protobuf on account of it not using reflection and generating more code for marshaling and unmarshaling. I gave it a whirl and it turns out to be way faster.

Using the same benchmark we made for #234 we get these results with the gogo-protobuf "gogo-fast" compiler:

Unmarshaling (which we do a lot of):
```
benchmark               old ns/op     new ns/op     delta
BenchmarkParseSSF-8     521           173           -66.79%

benchmark               old allocs     new allocs     delta
BenchmarkParseSSF-8     3              2              -33.33%

benchmark               old bytes     new bytes     delta
BenchmarkParseSSF-8     384           176           -54.17%
```

So it's 66% faster, and uses 54% less ram with one fewer alloc. Wow! I did not test marshaling since I don't know that we even do that in this code anywhere.

Marshaling (important for client libs):
```
benchmark                 old ns/op     new ns/op     delta
BenchmarkMarshalSSF-8     294           28.9          -90.17%

benchmark                 old allocs     new allocs     delta
BenchmarkMarshalSSF-8     1              0              -100.00%

benchmark                 old bytes     new bytes     delta
BenchmarkMarshalSSF-8     208           0             -100.00%
```

Well, that's exciting. 90% reduction and zero allocations. I suspect it's because we're not writing it it to anything in the test. We emit it to a `Writer` in the code.

#### Needs

I'm not certain how to add the gogo-protobuf binary to the `Dockerfile` so that this works with go-generate. I assume a `go get` added won't DWIW, or will it?

#### Test plan
gogo-protobuf says it is goprotobuf compatible, so I am relying on existing unit tests + a normal test cycle in QA.

r? @aditya-stripe