@ChimeraCoder 2017-08-28
We do marshal, within the `trace` package (which other Go services that use Veneur as a client library would use). Let's add a benchmark for that?