@ChimeraCoder 2017-08-28
Also, you'll have to update the build process in both Travis and and Jenkins. Updating the latter will fix the build error we're getting (though we have to update the former too, or else we're not properly testing the actual code we run).