@ChimeraCoder 2017-08-28
Maybe let's rebase this so `wtf cory` isn't in our git log forever? ;)

lgtm other than that!